Cleaned up Lapa script processing.
The script parser is still a central part of the Lapa processor, but in its new form the contract is the IScriptProcessor instead of a subset of different DI injected instances.

Still a processor needs all registrations of the parser, which in an ideal world should be refactored to an injection of a IScriptParser instance. Problem is that this currently isn't easy to facilitate.